+++ uefi/CHANGELOG.md
@@ -5,6 +5,8 @@
   the extended processor topology.
 
 ## Changed
+- **Breaking:** `proto::network::pxe::Server::server_type` and `Server::new`
+  now use `BootstrapType` instead of `u16`.
 - Made memory map types `#[repr(C)]`
 - Added `char16!` const-compatible macro as convenient replacement for `Char16::try_from().unwrap()`
 - `proto::debug::SystemContextARM` now contains the trailing `IFAR` field